+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/directconnectivity/TcpServerMock/TcpServer.java
@@ -36,7 +36,8 @@ import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.MultiThreadIoEventLoopGroup;
+import io.netty.channel.nio.NioIoHandler;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import io.netty.handler.logging.LogLevel;
@@ -60,8 +61,8 @@ public class TcpServer {
 
     public TcpServer(int port) {
         this.port = port;
-        this.parent = new NioEventLoopGroup();
-        this.child = new NioEventLoopGroup();
+        this.parent = new MultiThreadIoEventLoopGroup(NioIoHandler.newFactory());
+        this.child = new MultiThreadIoEventLoopGroup(NioIoHandler.newFactory());
         requestManager = new ServerRntbdRequestManager();
     }